cephadm: move file utility functions to file_utils.py
Pair-programmed-with: Adam King <adking@redhat.com>
